lib: Restructure BSP APIs
